# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Treinamento API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/luan/platforms": {
            "get": {
                "tags": [
                    "Luan Platforms"
                ],
                "summary": "List all LuanPlatforms",
                "operationId": "d110bd04a9ec882af30efae40c80f9be",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Luan Platforms"
                ],
                "summary": "Create a new LuanPlatform",
                "operationId": "337658b7791a5467a3c5e6251a531198",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string"
                                    },
                                    "icone": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/api/luan/platforms/{id}": {
            "get": {
                "tags": [
                    "Luan Platforms"
                ],
                "summary": "Get LuanPlatform by ID",
                "operationId": "f73cdeb7448f3d034c1aafae9a198d0b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Luan Platforms"
                ],
                "summary": "Update a LuanPlatform",
                "operationId": "dfda82a6c75e9b2028de8b56a45c97f3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string"
                                    },
                                    "icone": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated successfully"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Luan Platforms"
                ],
                "summary": "Delete a LuanPlatform",
                "operationId": "cba76cb77fc2abe39a3459d6b8a56fad",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted successfully"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/luan/reviews": {
            "get": {
                "tags": [
                    "Luan Reviews"
                ],
                "summary": "List all LuanReviews",
                "operationId": "4adfb23c09541e36bb35449bc952feb1",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Luan Reviews"
                ],
                "summary": "Create a new LuanReview",
                "operationId": "02a3aa7c922b979fc125ab64fa0a600b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "usuario_id": {
                                        "type": "string"
                                    },
                                    "jogo_id": {
                                        "type": "string"
                                    },
                                    "nota": {
                                        "type": "string"
                                    },
                                    "comentario": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/api/luan/reviews/{id}": {
            "get": {
                "tags": [
                    "Luan Reviews"
                ],
                "summary": "Get LuanReview by ID",
                "operationId": "630814f8ab03cc7fdef849d6f50f65d8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Luan Reviews"
                ],
                "summary": "Update a LuanReview",
                "operationId": "42044be6567044fd4f1bdf3378231bfc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "usuario_id": {
                                        "type": "string"
                                    },
                                    "jogo_id": {
                                        "type": "string"
                                    },
                                    "nota": {
                                        "type": "string"
                                    },
                                    "comentario": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated successfully"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Luan Reviews"
                ],
                "summary": "Delete a LuanReview",
                "operationId": "886ebdb3c15ce44d63d405079e3395f9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted successfully"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        }
    },
...
}